Extract options in RocksDBEngine, Scheduler, V8Server subfolders.
